pyproject.toml
[tool.poetry]
name = "spooky_numbers"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Pillow = "^9.2.0"
stability-sdk = "^0.2.6"
Unidecode = "^1.3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
